+++ packages/components/atoms/f-image-tile/src/components/ImageTile.vue
@@ -1,6 +1,9 @@
 <template>
     <div
-        :class="$style['c-imageTile']"
+        :class="[
+            $style['c-imageTile'], {
+                [$style['c-imageTile--selected']]: isToggleSelected
+            }]"
         data-test-id="image-tile-component">
         <a
             :class="[
@@ -21,14 +24,13 @@
             class="is-visuallyHidden"
             :class="$style['c-imageTile-checkbox']"
             data-test-id="image-tile-input"
+            :tabindex="!isLink ? 0 : false"
             @change="toggleFilter">
         <label
-            :class="[
-                $style['c-imageTile-label'], {
-                    [$style['c-imageTile-label--selected']]: isToggleSelected
-                }]"
+            :class="$style['c-imageTile-label']"
             :for="`imageTileToggle-${tileId}`"
-            data-test-id="image-tile-label">
+            data-test-id="image-tile-label"
+            :tabindex="!isLink ? -1 : false">
             <span
                 :class="$style['c-imageTile-imageContainer']"
                 :style="cssVars">
@@ -40,18 +42,26 @@
                     :alt="altText"
                     :role="isPresentationRole ? 'presentation' : false">
             </span>
-            <span :aria-hidden="isLink">
-                {{ displayText }}
+            <span
+                :class="$style['c-imageTile-textContainer']"
+                :aria-hidden="isLink">
+                <tick-icon :class="$style['c-imageTile-icon']" />
+                <span :class="$style['c-imageTile-text']">
+                    {{ displayText }}
+                </span>
             </span>
         </label>
     </div>
 </template>
 
 <script>
+import { TickIcon } from '@justeat/f-vue-icons';
 
 export default {
     name: 'ImageTile',
-    components: {},
+    components: {
+        TickIcon
+    },
     props: {
         href: {
             type: String,
@@ -149,13 +159,29 @@ export default {
 
 $image-tile-background-opacity: 0.7;
 $image-tile-background-color: $color-interactive-brand;
+$image-tile-selected: $color-content-positive;
+$image-tile-transition-duration: 0.2s;
+$image-tile-ease: ease-in-out;
+$image-tile-text-transform: translate3d(5px, 0, 0);
 
+@mixin image-tile-icon-selected-transform() {
+    opacity: 1;
+    transform: translate3d(0, 0, 0) scale(1) rotate(0);
+    width: 15px;
+}
+
+@mixin image-tile-focus() {
+    outline: 2px solid $color-focus;
+    border-radius: $radius-rounded-b;
+}
 
 .c-imageTile {
-    @include font-size(heading-m);
-
-    font-family: $font-family-base;
     position: relative;
+    width: 100%;
+
+    &:focus-within {
+        @include image-tile-focus();
+    }
 }
 
 .c-imageTile-link {
@@ -164,30 +190,38 @@ $image-tile-background-color: $color-interactive-brand;
     left: 0;
     right: 0;
     bottom: 0;
+
+    &:focus {
+        @include image-tile-focus();
+    }
 }
 
 .c-imageTile-link--toggle {
     display: block;
     position: static;
     pointer-events: none;
+
+    &:focus {
+        outline: none; // Focus styles not needed in toggle state.
+    }
 }
 
 .c-imageTile-label {
     display: flex;
     flex-flow: column wrap;
-}
 
-.c-imageTile-label--selected {
-  border: 1px solid green;
+    &:focus {
+        outline: none; // Prevents Safari doubling focus styles.
+    }
 }
 
 .c-imageTile-imageContainer {
-  display: block;
-  background-color: rgba($image-tile-background-color, $image-tile-background-opacity);
-  border-radius: $radius-rounded-b;
-  background-image: var(--bg-image);
-  padding-top: (3 / 5) * 100%; // 5:3 aspect ratio
-  position: relative;
+    background-color: rgba($image-tile-background-color, $image-tile-background-opacity);
+    border-radius: $radius-rounded-b;
+    background-image: var(--bg-image);
+    display: block;
+    padding-top: (3 / 5) * 100%; // 5:3 aspect ratio
+    position: relative;
 }
 
 .c-imageTile-image {
@@ -199,5 +233,70 @@ $image-tile-background-color: $color-interactive-brand;
     height: 100%;
 }
 
+.c-imageTile-textContainer {
+    margin-top: spacing(x2);
+    display: flex;
+    max-width: 100%;
+
+    .c-imageTile--selected & {
+        color: $image-tile-selected;
+    }
+}
+
+.c-imageTile-icon {
+    align-self: center;
+    opacity: 0;
+    transform: translate3d(-10px, 0, 0) scale(0.5) rotate(-60deg);
+    width: 0;
+    will-change: transform, opacity;
+
+    @media (prefers-reduced-motion: no-preference) {
+        transition: transform $image-tile-transition-duration $image-tile-ease,
+                    opacity $image-tile-transition-duration $image-tile-ease,
+                    width $image-tile-transition-duration $image-tile-ease;
+    }
+
+    .c-imageTile--selected & {
+        @include image-tile-icon-selected-transform();
+    }
+
+    .c-imageTile--selected & path {
+        fill: $image-tile-selected;
+    }
+
+    @include media('>=mid') {
+        .c-imageTile:hover & {
+            @include image-tile-icon-selected-transform();
+        }
+    }
+}
+
+.c-imageTile-text {
+    @include font-size(body-s);
+    display: block;
+    font-family: $font-family-base;
+    font-weight: $font-weight-regular;
+    margin-right: spacing(x2);
+    overflow: hidden;
+    text-overflow: ellipsis;
+    transform: translate3d(0, 0, 0);
+    white-space: nowrap;
+    width: 100%;
+    will-change: transform;
+
+    @media (prefers-reduced-motion: no-preference) {
+        transition: transform $image-tile-transition-duration $image-tile-ease;
+    }
+
+    .c-imageTile--selected & {
+        transform: $image-tile-text-transform;
+    }
+
+    @include media('>=mid') {
+        .c-imageTile:hover & {
+            transform: $image-tile-text-transform;
+        }
+    }
+}
 </style>