+++ packages/components/atoms/f-image-tile/CHANGELOG.md
@@ -3,6 +3,23 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+v0.5.0
+------------------------------
+*January 28, 2022*
+
+### Added
+- Focus styles for keyboard navigation
+
+v0.4.0
+------------------------------
+*January 26, 2022*
+
+### Added
+- Styled and truncated text
+- Transitions accounting for prefers-reduced-motion
+
+
 v0.3.0
 ------------------------------
 *January 19, 2022*